pyproject.toml
[tool.poetry]
name = "qa-guru-python-5-24"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "qa_guru_python_5_24"}]

[tool.poetry.dependencies]
python = "^3.11"
selenium = "4.5.0"
selene = {version = "^2.0.0-beta.17", allow-prereleases = true}
Appium-Python-Client = "^2.8.1"
pytest = "^7.3.2"
pytest-xdist = "^2.5.0"
allure-pytest = "^2.10.0"
requests = "^2.28.1"
pydantic = "^1.10.5"
python-dotenv = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
